Moving PID Controller to its own file
