+++ prototype/pid.h
@@ -1,0 +1,12 @@
+/*
+ * Sierra College Computer Engineering Club: Sparkfun AVC
+ * PID Controller
+ * Authors: Kristoffer Semelka
+ */
+
+#ifndef _PID_H_
+#define _PID_H_
+
+float pidAdjust(float headingDiff);
+
+#endif
